[project] Move conn logger just to bootnode
