+++ cmd/harmony/main.go
@@ -101,9 +101,7 @@ var (
 	devnetNumShards   = flag.Uint("dn_num_shards", 2, "number of shards for -network_type=devnet (default: 2)")
 	devnetShardSize   = flag.Int("dn_shard_size", 10, "number of nodes per shard for -network_type=devnet (default 10)")
 	devnetHarmonySize = flag.Int("dn_hmy_size", -1, "number of Harmony-operated nodes per shard for -network_type=devnet; negative (default) means equal to -dn_shard_size")
-	// logConn logs incoming/outgoing connections
-	logConn     = flag.Bool("log_conn", false, "log incoming/outgoing connections")
-	keystoreDir = flag.String("keystore", hmykey.DefaultKeyStoreDir, "The default keystore directory")
+	keystoreDir       = flag.String("keystore", hmykey.DefaultKeyStoreDir, "The default keystore directory")
 	// logging verbosity
 	verbosity = flag.Int("verbosity", 5, "Logging verbosity: 0=silent, 1=error, 2=warn, 3=info, 4=debug, 5=detail (default: 5)")
 	// dbDir is the database directory.
@@ -416,9 +414,6 @@ func createGlobalConfig() (*nodeconfig.ConfigType, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot create P2P network host")
 	}
-	if *logConn && nodeConfig.GetNetworkType() != nodeconfig.Mainnet {
-		myHost.GetP2PHost().Network().Notify(utils.NewConnLogger(utils.GetLogger()))
-	}
 
 	nodeConfig.DBDir = *dbDir
 
@@ -604,7 +599,6 @@ func setupViperConfig() {
 	viperconfig.ResetConfUInt(devnetNumShards, envViper, configFileViper, "", "dn_num_shards")
 	viperconfig.ResetConfInt(devnetShardSize, envViper, configFileViper, "", "dn_shard_size")
 	viperconfig.ResetConfInt(devnetHarmonySize, envViper, configFileViper, "", "dn_hmy_size")
-	viperconfig.ResetConfBool(logConn, envViper, configFileViper, "", "log_conn")
 	viperconfig.ResetConfString(keystoreDir, envViper, configFileViper, "", "keystore")
 	viperconfig.ResetConfInt(verbosity, envViper, configFileViper, "", "verbosity")
 	viperconfig.ResetConfString(dbDir, envViper, configFileViper, "", "db_dir")
